+++ public/js/voice.js
@@ -67,8 +67,8 @@ class VoiceManager {
 
     this.rtcConfig = {
       iceServers: [
-        { urls: 'stun:stun.l.google.com:19302' },
-        { urls: 'stun:stun1.l.google.com:19302' }
+        { urls: 'stun:stun.stunprotocol.org:3478' },
+        { urls: 'stun:stun.nextcloud.com:3478' }
       ]
     };